Deploy swiftly-install 0.3.0
This release modifies swiftly-install's default behavior when encountering an
existing installation of swiftly to updating only the swiftly binary rather than
performing a clean installation, which was the previous default. The
previous behavior can be achived by passing in the --overwrite flag.

This release also fixes a bug where swiftly-install would not clean up
any existing symlinks when overwriting an existing swiftly installation.