macOS support for swiftly
It works much like it does already for Linux with some notable
differences:

  * The toolchains are installed using the pkg files and macOS
    installer
  * The toolchain directory is ~/Library/Developer/Toolchains
    instead of ~/.local/share/swiftly/toolchains

Create a MacOS struct that implements the existing Platform
protocol. Make a platform-specific target for this module. Bump
the required swift toolchain version to resolve compiler errors
and set the minimum macOS version to 13.

Update the README.md with some macOS details and fix some of the
details that were outdated for Linux.

Since macOS configuration is very simple compared to Linux, the
configuration can be auto-created from Swift directly if it is
missing.

Add some helpful notes regarding the need to rehash the zsh on
macOS since even when the swiftly bin directory has higher
precedence in the PATH it sometimes gets snagged on the
/usr/bin/swift, which doesn't detect the user installed toolchains
and sometimes tries to get the user to install Xcode.

Make the shell script swiftly installer capable of operating in
a standard macOS environment. First, detect that the environment
is macOS, and then adjust the getopts for macOS's more limited
implementation with the short opts. Also, remove any of the Linux
specific steps to detect the distribution, check for gpg, and
attempt to install Linux system packages.