Swiftly self-install when running the binary
A possible mode of swiftly install is to download the binary from
a trusted source and then begin running it. When it is first run
in this way it should auto-detect if it is being run outside of
an installation and go through the process of installing itself
in the user's account.

Copy the installation and configuration logic into swiftly:
* Check if downloads from github/swift.org can work with trusted
  CA certificates
* Check if GPG is installed (Linux)
* Prompt to install swiftly with the proposed directory structure
* Provide a set of overrides for some of the install behaviour
  such as assume yes to any prompts, not updating the user's profile,
  and overwrite the swiftly that is installed with the current one
* Auto-detect the Linux distribution from the supported list,
  or prompt if this cannot be determined
* Import the gpg keys from swift.org into the keychain if it isn't
  there already (Linux)
* User shell detection (bash, zsh, or fish)

If the user continues to run the downloaded swiftly binary instead
of the installed version provide a helpful message to show them
how to source the environment file to begin using the installed
version. This can happen if they continue using the same shell
session where the install was performed.