Remove the swiftly-install.sh script entirely
Since swiftly is capable of self-installation there's no need for
the shell script. Users will download a swiftly binary and use it
to accomplish the same tasks.

Show the command that the user should install on their system to
support a particular toolchain that they've installed. This
replaces the previous mechanism of the shell script that would
prompt the user and install the packages right away. Note that
this new mechanism checks each package and displays only the
missing packages.

Provide a platform global option that allows the user or some
automation to pick the correct platform, bypassing the auto
detection mechanisms on Linux.

Make the overwrite option more closely match what the shell script
would do, which is to clear out directories that were there from
a previous swiftly install.

Refactor the install test scripts to use a built version of swiftly
instead of the shell script. Also, remove the test cases and checks
for the installation of system packages. The new approach will
prompt the user with the list of system packages that are expected
to be installed. Plus, swiftly itself will check if the tools it
needs to be installed are present and abort if they aren't availanble.

Update the shell tests so that they update the user's login shell
rather than relying on the SHELL environment variable. Since swiftly
is unlikely to have access to the parent shell's SHELL environment
variable it has checks for the user's login shell setting.

Update the install test scripts so that they can function on other
supported Linux distributions such as ubi9.

Add instructions on how to run the different test suites.