Proxy support for swiftly
Instead of providing symbolic links that are globally target
a particular toolchain that is 'used' swiftly acts as a proxy
for each of the binaries in a toolchain. This allows swiftly to
use a rich set of configurations to determine on each tool
invocation to determine what toolchain should be used. For
example, special command-line arguments like '+5.10.0' can be
provided to any tool and the proxy will take that as the toolchain
selector for that call. If the toolchain isn't installed then
swiftly can automatically download and install it for you.

Have swiftly install symlinks for the tools in a toolchain such
as 'swift', 'swiftc', and other allow-listed tools for the proxy
targeting swiftly itself. Have swiftly detect if the first command
line argument, argv[0], is not swiftly itself and then use its
proxy mechanism to find an installed toolchain to route the
invocation.

When proxying the highest precedent for the toolchain selection is
a special argument with the syntax '+<selector>'. Swiftly will use
this to determine if there is already a suitable toolchain installed.
If no matching toolchain is installed then swiftly will proceed to
install it. Once the toolchain is installed swiftly will proxy to
the tool using all of the other command line arguments.

The next highest precedence is the presence of a '.swift-version'
in the current working directory, or any ancestor directory. If
the file contains a toolchain selector this will be used following
the same process described previously.

If no other toolchain selector is available then swiftly will default
to the toolchain that is in use in the swiftly global config.json of
the installation. This serves as the global default toolchain.

Update swiftly use so that it can it updates a swift version file if
one is present to show which toolchain selector is in use for the
current package. The fallback is the global inuse configuration.
When using the use command to use a different installed toolchain
swiftly will update the swift version file with the selected toolchain
or will attempt to discover the root of the swift package and create
a new file in there unless instructed by the new global flag that
forces it to update the global configuration.

Add a new swiftly init subcommand that does a local installation of
swiftly into the user's environment. This contrasts with the previous
approach where any subcommand invocation could cause an installation
to happen. Now the subcommand invocations just validate that they
operate in a valid swiftly installation and abort if they are not.

When a user downloads swiftly from the official website. It may
come in the form of a system package or be a bare download. In either
case there should be a mechanism to trigger a user level installation
to happen with the init subcommand. The proxy mechanism is based on
the executable name to route the invocation to a specific path. If
the swiftly executable is invoked as 'swiftly-init' as it can be
named in the system package (e.g. macOS pkg) or in the direct download
from the official website then this will route directly to the init
subcommand to begin the user installation process. Any extra command
line flags, such as overwrite, and the platform hint for Linux will
be sent along.

Create a shell script to create a swiftly universal binary in a
macOS package with the swiftly-init name. The package will place
swiftly-init in /usr/local/bin, which is often on the user's path
at the terminal so that they can run it immediately to do the user
install.