Update usage of unsafe API, upgrade libarchive to 3.7.4, upgrade async http client (#120)
Memory safety and security fixes

Reduce the use of unsafe swift constructs in favour of ones that
are memory safe. In the HTTP client remove the use of the unsafe
mutable pointer in favour of direct Data conversion before writing
the downloaded file to disk.

Upgrade the version of libarchive to 3.7.4, which includes some
fixes.

Fix the changes to the SwiftlyHTTPClient that changed it to
use the shared async HTTPClient. That broke the interactions with
the GitHub REST API's. Update to the new async http client that
fixes problems with decompression limits.

Co-authored-by: Chris (SPG) McGee <christie_mcgee@apple.com>