DRAFT: Merge installation logic into swiftly as an init subcommand
Instead of relying on a separate shell script with certain issues
swiftly can perform its own user initialization with many of the
features of the script, including prompts, flags, and options. It
also prepares swiftly for possible delivery through system package
managers.

Provide a design for the new installation system in the DESIGN.md.

Change SwiftlyVersion to make it more conformant to the semantic
version specification (https://semver.org) with respect to pre
releases. Add test cases that cover typical scenarios. Update the
swiftly version to a "dev" pre-release.