DRAFT: Vendor libarchive with SwiftPM
Use SwiftPM to compile and link in libarchive using a vendored
version of it. This will help to ensure that it is statically linked
into the final binary in both the dev environment and at release.
Also, it removes the need for a separate C compiler toolchain to
be present when producing a released binary, just the same swift
toolchain that is being used to make the swiftly binary.