Add DocC documentation for swiftly
Create a plugin and tool that can generate a CLI reference for swiftly
and generate the current latest document.

Organize the help into a GSG (Getting Started Guide), a selection
of HOWTOS, and Reference topics. Write an initial version of each.

Create a Swift Package Index metadata file so that swiftly can
be submitted there once the documentation is ready.

Update the RELEASING.md file to check the documentation sanity and
command-line reference, regenerating it if necessary.