macOS support for swiftly (#121)
macOS support for swiftly

It works much like it does already for Linux with some notable
differences:

  * The toolchains are installed using the pkg files and macOS
    installer
  * The toolchain directory is ~/Library/Developer/Toolchains
    instead of ~/.local/share/swiftly/toolchains
  * The swiftly shared directory is
    ~/Library/Application Support/swiftly as it this is a
    more typical place for macOS applications to store their
    supporting files

Create a MacOS struct that implements the existing Platform
protocol. Make a platform-specific target for this module. Bump
the required swift toolchain version to resolve compiler errors
and set the minimum macOS version to 13.

Update the README.md with some macOS details and fix some of the
details that were outdated, both there and in DESIGN.md.

Add some helpful notes regarding the need to rehash the zsh on
macOS since even when the swiftly bin directory has higher
precedence in the PATH it sometimes gets snagged on the
/usr/bin/swift, which doesn't detect the user installed toolchains
and sometimes tries to get the user to install Xcode.

Make the shell script swiftly installer capable of operating in
a standard macOS environment. First, detect that the environment
is macOS, and then adjust the getopts for macOS's more limited
implementation with the short opts. Also, remove any of the Linux
specific steps to detect the distribution, check for gpg, and
attempt to install Linux system packages.

Add support for macOS CI. Read environment variables for a possible
HTTP proxy, and use it for the http client when running the tests.
Refactor the mechanisms used to override the http client and the
request executor to support proxies, while still supporting the mock
of both toolchain downloads and using the request handler lambda.