Merge installation logic into swiftly as an init subcommand (#127)
Merge installation logic into swiftly as an init subcommand

Instead of relying on a separate shell script with certain issues
swiftly can perform its own user initialization with many of the
features of the script, including prompts, flags, and options. It
also prepares swiftly for possible delivery through system package
managers.

Provide a design for the new installation system in the DESIGN.md.

Bump the Swiftly version to 0.4.0 with the dev suffix to signal a
new development version of swiftly that can be compiled and used
to install for early adoption and testing.

Create a new init subcommand that performs the post-download
init procedure for the user. Perform a validation whether swiftly is
initialized at the beginning of each other subcommand and also
check its version. The version is now encoded in the config.json file
in order to detect upgrade or attempted downgrade.

Update the install command so that it is capable of reporting missing
system packages, providing the command that the user can run in
order to install them. Add a special post installation file option that
will provide the command in a file that can be run after installation
to install the system packages. Add post installation script generation
options to Platform protocol. Add platform (aka. distribution) detection
capabilities to the Linux platform, and create the package lists that are
needed for each distribution. Use the distribution-specific list and
detection commands to decide if the packages are installed in the
user's system so that a tailored post-installation script is generated,
or none at all if the user's system is ready for the toolchain.

Create a HOWTO guide to demonstrate how to install swiftly and
toolchains unattended in places like CI systems. Highlight the various
concerns that should be considered in the process of automating it.

Since the platform definition auto detection is added to the init
subcommand update the test harness so that it is capable of auto
detecting the platform instead of relying on the environment variables.
The environment variables remain as an override for testing purposes.