Use swift.org API for getting releases and some snapshot toolchains (#153)
See issue #149

Using the new swift.org API is more accurate than the GitHub API, also it isn't subject to the same rate limits. Use the swift.org API instead of the GitHub API for fetching metadata about the released swift toolchains, and the newer swift branches for the snapshots.

However, there is currently no metadata for snapshots on older branches than 6.0 or main. In this case use the GitHub API to get the metadata, but only for swiftly install and not the swiftly list-available commands, since the latter is very expensive when using the GitHub API.