Integrate SwiftFormat into the swiftly project dependencies (#158)
The SwiftFormat that is run in the CI is a very specific version
that is unlikely to be the one a dev has installed with homebrew
or discovered for themselves and built from source. Discovering
the version involves finding the magic in the lint.dockerfile.

SwiftPM provides a mechanism for bringing in product dependencies
so that you can simply run using the toolchain. The dependencies
can be specified using an exact match on a particular version.

Add the SwiftFormat dependency as a package dependency in swiftly
at the version that is currently used in the soundness CI checks.
Use the the swift command to invoke the 'swiftformat' product from
the dependency instead of manually git cloning and building it
manually in the dockerfile.

Provide a rationale for the dependency in the Package.swift for
anyone who is curious about it.