Enhancements to the release build script
Add macOS support for making releases there.

Rewrite the release builder script in Swift for better reproducibility
and ergonomics. It runs using this command:

    swift run build-swiftly-release

Perform necessary system checks up front so that anyone running the
script can make any necessary adjustments so that the release can be
built.

Print the location of the final release product file at the end for
easy of finding the path to copy to the final release destination.