Enhancements to the release build script (#165)
Add macOS support for making releases there.

Rewrite the release builder script in Swift for better reproducibility
and ergonomics. It runs using this command:

    swift run build-swiftly-release

Perform necessary system checks up front so that anyone running the
script can make any necessary adjustments so that the release can be
built.

Print the location of the final release product file at the end for
easy of finding the path to copy to the final release destination.

Add license file for both Linux and macOS.

Update the releasing document with the new release instructions.

Update the self-update logic to work with new the release artifacts and a proposed swift.org API.

Import GPG keys on every install to get new signing keys from swift.org.

Fix all of the swift.org urls so that they use www.swift.org to avoid redirection.