Add signature cert and package identifier parameters to package script (#174)
When building macOS packages with pkgbuild, a signing certificate
needs to be provided up-front, as is the package identifier. Change
the package script so that it can accept the location of the certificate
and override the package identifier.

Add a note in the RELEASING.md about the --cert option for the
release builder.