Fix package install detection for packages that are partially installed (#178)
Fix package install detection for packages that are partially installed.

On Debian/Ubuntu Linux that use the dpkg/apt package manager there
is a possibility that a package is not in a fully installed state.
It can be removed, leaving its configuration files, or the install
could have failed. Check rigorously that the package is fully installed
without error before producing the command-line.

Also, correct the install command to use apt-get instead of apt since
that command-line is guaranteed to be stable.