Add post-merge CI and build release workflows (#185)
After squashing and mergin a commit on top of what's on the main
branch there can be merge issues. Add a post-merge (ie. push)
trigger for the same workflows that are used for pull requests to
help flag merge issues.

For these workflows, it will be useful to upload the release
artifacts in case someone would like to manually test it. In the
future there is also a possibility of larger e2e test suites that
could run on fresh Linux containers without existing swift toolchains.

Add a new workflow to produce release artifacts for Linux on a
manual workflow dispatch either from the GitHub UI, or via the REST
interface.