Test swiftly release artifacts in a plain Linux environment
Currently, there is no test coverage of a swiftly installation
where there is no swift support at all, and nothing involving
swiftly. Add a github workflow that performs a swiftly install
using the release artifact.

In order to facilitate this create an E2E test that performs
these steps, but only in an environment where it can mutate the
system. Also, make this test deployable to a plain Linux system
with no swift support installed on it.