Download toolchains using proxy if one is set in the environment (#189)
Swiftly doesn't currently handle the standard HTTP proxy server environment variables when making its network connections. It only handles it as a necessity in the current macOS CI environment and only in the tests. Make the built-in http executor read these environment variables if they are present and use those for all http requests. Add documentation in the GSG showing how to use this feature.