Add documentation related build and checks to GitHub workflows (#186)
There's nothing in the pull request verification that checks if
the command-line references are up-to-date. Add a job to the pull
request update workflow that verifies this.

There's nothing in the release build workflow that generates the
documentation and uploads the documentation artifacts so that they
can be published. Add a step to the build release job that generates
the documentation so that it can be published to swift.org.

Overcome git dubious repo error.

Give the uploaded artifacts unique names to avoid collisions.