Streamline the swiftly init process (#177)
The init process just installs swiftly itself at the moment. Most
users will immediately install a swift toolchain, most likely the
latest available one. On Linux, there's confusing gpg messages that
most users don't need to be aware.

The init process will provide a summary of things that are going
to happen to the user's system, including the addition of GnuPG keys
on Linux, and the installation of the latest swift toolchain so that
they can agree, or abort the entire process. When the process runs
the user is given line-level and high level processes, not internal
details. Add a verbose mode for more details, such as the messages
that come from GnuPG on Linux.

Add an option to the init subcommand to allow swiftly to be
installed without the latest available swift toolchain so that
advanced users can decide how to install a toolchain themselves
after the swiftly installation.

Update the documentation with the more automated workflow.

Add more detail to the getting started guide.

Rework the getting started guide to have a platform selector for
linux/macOS and hide platform-specific details

Permit bare swiftly command to start the init workflow when not installed.

Quiet the macOS installer messages behind the verbose flag.

Provide more verbose untarring messages in Linux behind the verbose flag.