Adopt the static linux sdk (#196)
With the Static Linux SDK swiftly can be compiled on a particular
Linux and processor architecture and the binary can run on a very
wide variety of versions and distributions. This eliminates the need
to build swiftly on "older" Linux distributions to produce a binary
that runs on the "newer" ones>

Bump the version of swift-tools-support-core so that it compiles
successfully in a Static SDK environment with Musl.

Fix the readLine function to add a newline to the prompt so that it
prints it out before waiting on user input on stdin.

Update the build-swiftly-release tool. First, detect the version of
swift so that the matching version of the Static SDK is downloaded
and installed into the toolchain. Use the Musl that is inside it to
configure and compile libarchive along with a special must-clang
wrapper script that sets the correct compile options to use Musl
instead of the system's glibc. Finally, build the swiftly binary with
the Static SDK / Musl linking it together with the Musl compiled
libarchive.

Deprecate the old flag that would decide between RHEL UBI9
or Amazon Linux 2 as the required release building platform and
eliminate the supported Linux check.