Add support for Debian 12 in the install script
This makes the relatively mechanical changes needed to add support for Debian 12 in the install script. Debian Docker nightlies are not yet available, so we use Ubuntu's Dockerfile for computing the dependency list for now.

This has been tested on a Debian 12 installation and is able to correctly install swiftly.