Add ability to temporarily disable swiftly
Adds two new commands, `swiftly unlink` and `swiftly link`, which
will disable and reenable swiftly's management of the active toolchain.

The `unlink` command removes the symlinks to toolchain binaries in the
swiftly bin directory that is in the user's path. This allows the rest
of the `$PATH` to be searched for available toolchain installations,
falling back to the system default. On macOS with Xcode installed this
has the effect of falling back to the toolchain in the user's installed
Xcode.

The `link` command reinstates the symlinks to the `inUse` toolchain,
which allows swiftly to resume management of the active toolchain.