Add a modeled command-line for better checking (#330)
Swiftly has many different commands that it invokes in different
parts of its code base, supporting tooling, and tests. This list
includes tar, git, pkgutil, and dscl among oseveral others.

For the most part these tools are called using arrays that are
manually assembled at the call site. In order to exercise that
callsite in tests the test must necessarily become "medium" size
because it involves integration testing with that tool, instead of
"small", which is where most tests should be. A small test of that
array would verify that the length has the correct arguments,
number, spelling, and order. Also, there are various types of string
interpolation in these arrays, which limit the type-safety.

Introduce a new way of modeling command-line invocations using Swift
types. Commands and subcommands are modeled using structs. Use
functions to assemble the structs, and methods to assemble structs
for subcommands. Arguments become function arguments, with default
values if they are optional. Flags, and options will be described
using enum cases with carrying values for options.

Some commands are runnable, while others are not. Introduce a
Runnable protocol that marks runnable commands, and provides a
default implementation of a run() method to run the assembled command.

Other commands can produce output. Introduce an Output protocol and
default implementation that runs the command and receives the output
as a string that the callsite can use.

Some callsites make decisions based on whether a tool is present
or not. Provide a convenient way to verify whether a command exists
on the system. Use this mechanism to skip tests when they are run
in an environment that doesn't have the tool.

Pick one command, dscl, to use with the new modeled command-line.
Write small tests that don't invoke the tool. Instead they
verify the configuration that is produced with different typical
inputs from swiftly. Add a single medium size test that runs dscl
from a modeled command, enabled only if dscl exists on the system.

Add tags for medium, and large tests. Mark existing large tests,
which is the HTTPClientTests. Leave all of the remaining tests as
untagged, (i.e. small). Mark the CommandLineTest that invokes the
dscl tool as medium.