Model the pkgbuild command
Model the pkgbuild command with the usual structure used for other
commands.

Provide a custom string convertible implementation for configuration
so that the configuration can be easily converted into a string for
logging and comparison purposes.

Create a runEcho() implementation for the build swiftly release
script so that commands can output their command-line to the log
for reproducibility.