Model additional system commands (#332)
Model the pkgbuild command with the usual structure used for other
commands.

Provide a custom string convertible implementation for configuration
so that the configuration can be easily converted into a string for
logging and comparison purposes.

Create a runEcho() implementation for the build swiftly release
script so that commands can output their command-line to the log
for reproducibility. Add custom string capabilities to Configuration
and Arguments to support this.

Model the Linux getent command.

Model a small subset of the git command.

Replace array += with append to save on allocations.