Build swiftly release cleanup (#335)
Remove the copies of bits of SwiftlyCore functionality in the build-swiftly-release script:

runProgramEnv|Output
SwiftlyRelease/SwiftlyPlatform (for JSON decoding from the REST endpoint)
curl
Remove the curl dependency from the release script using the AsyncHTTPClient instead
to download libarchive.

Migrate the release script from URLs and FileManager to the new FilePath and FileSystem
component for improved readability, and type safety.