Tests: Disable Git Commit Signing in Tests
The test suite creates a tiny test suite with a custom user email.
If a developer working on Swiftly has commit signing enabled, that tool
will see that the committer email and the signing identity don't match,
and will fail to sign the commit, causing the `testGit` test to fail.

Disabling gpg signing commits in the temporary test repo.