Update UseTests: Swallow unknown version errors
Now that the mock home directory isn't swallowing function errors, these
are resulting in test failures when they aren't actually intended.
In these cases, we use `#expect` to verify the version, but swallow any
errors returned by the swiftly invocation.

In all cases, these were the result of trying to switch to a toolchain
that didn't exist. This should emit an error message, but leave the
selected toolchain unchanged.