Follow symlinks when validating Swiftly is linked
When swiftly is installed via homebrew it is placed in
`/opt/homebrew/bin/swiftly`, which is a symlink to a location like
`/opt/homebrew/bin/Cellar/swiftly/1.1.0/bin/swiftly`.

When checking if Swiftly is linked we check to see if the `swiftly`
symlink in the swiftly bin directory is pointing to the swiftly
executable. However with Homebrew, the symlink is not pointing to the
swiftly executable but instead to _another symlink_ that in turn points
to the executable.

Augment `fs.readlink` to accept a `follow` paramter, which will follow
symlinks until we reach the real underlying file.

By following the trail of symlinks from the swiftly bin directory to the
underlying swiftly executable the isLinked check will succeed and no
longer report that swiftly is unlinked when its actually linked
correctly.