Don't install the SL SDK into the user account
When building the release binary for swiftly it uses the static
linux SDK with musl to create a system independent binary. The
SDK was being installed by the build-swiftly-release tool into
the user's home directory with "swift sdk install" command and
it tries to uninstall it afterwards. This can pollute a user's
system or remove the SDK if they installed it already.

In this approach the SDK is extracted into a scratch directory
and that directory is provided to the "swift build" command so
that it can find it there.